requirements.txt
torch>1.0
Pillow<7.0.0
tqdm
scikit-learn
PyWavelets~=1.1.1
matplotlib
# Development dependencies
cython>=0.29